refactor: dashboard translations import
